Implementacion Actividad finalizada de Editar Perfil

+++ app/src/main/java/com/example/rent4xalapa/poko/Usuario.kt
@@ -1,10 +1,10 @@
 package com.example.rent4xalapa.poko
 
 data class Usuario(
-    val idUsuario: Long,
+    val idUsuario: Int,
     val nombre: String,
     val correo: String,
     val contrasena: String,
-    val telefono: Int,
+    val telefono: Long,
     val ine: String
 )
